Fixes bug with callback on data_layers model
The callback is used to defined the value of is_ready for the
data_layer.

But when that value is false the callback's return value is also false,
which causes the record to not be saved. The return value needs to be
true always, regardless of the value of is_ready.