Fixes duplicated pages on public interface.
Removes includes method from the controller as in this case N is 1 so
we do not need to use includes to avoid issues
with N+1 queries being triggered.

This also fixes a bug where duplicated pages were being fetched because
of the lack of distinct on the valid_pages association.

Another reason to remove includes from the controller is related with a
bug in ActiveRecord whereby the includes method strips group and other closes
from associations it tries to include. This bug is currently stale. See
issue: https://github.com/rails/rails/issues/15854